compatible processing of screen size change
New preference allows user to chose when terminal to calculate screen size.

Update of terminal size in onSizeChaged() event works well when terminal is run in "window" mode.
In this mode calculation in onMeasure() based on "WindowVisibleDisplayFrame" and "GlobalVisibleRect" produces wrong results
(see https://gitlab.com/termapps/termoneplus/issues/10).

So onSizeChaged() works well with most of soft-keyboards(IME method).
Problem is that in some cases onSizeChaged() is not called if orientation is "landscape".
It depend of Android version and IME method.

Remark: In "full-screen" mode is restored use onMeasure() but with more simple implementation.

Resolves issue #17.